+++ test/egyptian_test.dart
@@ -30,17 +30,6 @@ void main() {
       },
     );
 
-    test('Making sure that static methods can be accessed', () {
-      expect(
-        () {
-          EgyptianFraction.clearCache();
-
-          return EgyptianFraction.cachingEnabled;
-        }(),
-        isTrue,
-      );
-    });
-
     test('Making sure that caching properly works', () {
       final egyptian = EgyptianFraction(fraction: Fraction(2, 5));
 
@@ -118,7 +107,6 @@ void main() {
 
       // Calling 'toString' twice to make sure that CI coverage also covers the
       // case where the instance is cached
-      EgyptianFraction.clearCache();
       expect('$egyptian', equals('1/2 + 1/8'));
     });